[FEAT] 문제 및 답안 조회 기능 구현

+++ src/main/java/com/example/ai_tutor/domain/note/domain/Note.java
@@ -28,7 +28,7 @@ public class Note extends BaseEntity {
     private String title;
 
     // 제한시간
-    private int limitTime;
+    private long limitTime;
 
     // 마감시간
     private LocalDateTime endDate;
@@ -54,7 +54,7 @@ public class Note extends BaseEntity {
 //    }
 
     @Builder
-    public Note(Folder folder, String title, int limitTime, LocalDateTime endDate, String code){
+    public Note(Folder folder, String title, long limitTime, LocalDateTime endDate, String code){
         this.folder = folder;
         this.title = title;
         this.limitTime = limitTime;
